fix: 승리자 출력 함수 수정
