+++ NumberBaseball/NumberBaseball/validateFunction.swift
@@ -31,17 +31,3 @@ func isNumberFromOneToNine(numbers: [String]) -> Bool {
     return numbers.count == numbersFromOneToNine.count ? true : false
 }
 
-
-func initializeGame() {
-    leftCount = 9
-    isUserWin = false
-    computerNumbers = makeThreeNumbers()
-}
-
-func printMenu() {
-    print("""
-          1. 게임시작
-          2. 게임종료
-          원하는 기능을 선택해주세요 :
-          """, terminator:" ")
-}
